node: extract node rest api handler from daemon
Currently, the (cluster)node rest api handler is implemented as part of the
daemon itself and is buried in the `status.go` file.

To further minimize the number of dependencies on the daemon, this commit
extracts the node rest api handler into the nodemanager hive cell in package
`pkg/node/nodemanager`.
