+++ pkg/node/manager/rest_api_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package manager
 
 import (
 	"testing"
@@ -13,16 +13,13 @@ import (
 
 	"github.com/cilium/cilium/api/v1/models"
 	. "github.com/cilium/cilium/api/v1/server/restapi/daemon"
-	"github.com/cilium/cilium/daemon/cmd/cni/fake"
 	fakeTypes "github.com/cilium/cilium/pkg/datapath/fake/types"
-	"github.com/cilium/cilium/pkg/node/manager"
 	nodeTypes "github.com/cilium/cilium/pkg/node/types"
-	"github.com/cilium/cilium/pkg/nodediscovery"
 	"github.com/cilium/cilium/pkg/option"
 )
 
 type GetNodesSuite struct {
-	nm manager.NodeManager
+	nm NodeManager
 }
 
 var fakeConfig = &option.DaemonConfig{
@@ -32,11 +29,11 @@ var fakeConfig = &option.DaemonConfig{
 }
 
 func setupGetNodesSuite(tb testing.TB) *GetNodesSuite {
-	option.Config.IPv4ServiceRange = AutoCIDR
-	option.Config.IPv6ServiceRange = AutoCIDR
+	option.Config.IPv4ServiceRange = "auto"
+	option.Config.IPv6ServiceRange = "auto"
 
 	h, _ := cell.NewSimpleHealth()
-	nm, err := manager.New(fakeConfig, nil, &fakeTypes.IPSet{}, nil, manager.NewNodeMetrics(), h)
+	nm, err := New(fakeConfig, nil, &fakeTypes.IPSet{}, nil, NewNodeMetrics(), h)
 	require.Nil(tb, err)
 
 	g := &GetNodesSuite{
@@ -59,9 +56,9 @@ func Test_getNodesHandle(t *testing.T) {
 
 	var zero int64
 	type args struct {
-		params  GetClusterNodesParams
-		clients map[int64]*clusterNodesClient
-		daemon  *Daemon
+		params      GetClusterNodesParams
+		clients     map[int64]*clusterNodesClient
+		nodeManager NodeManager
 	}
 	type want struct {
 		clients   map[int64]*clusterNodesClient
@@ -75,15 +72,12 @@ func Test_getNodesHandle(t *testing.T) {
 		{
 			name: "create a client ID and store it locally",
 			setupArgs: func() args {
-				nodeDiscovery := nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil)
 				return args{
 					params: GetClusterNodesParams{
 						ClientID: &zero,
 					},
-					daemon: &Daemon{
-						nodeDiscovery: nodeDiscovery,
-					},
-					clients: map[int64]*clusterNodesClient{},
+					nodeManager: g.nm,
+					clients:     map[int64]*clusterNodesClient{},
 				}
 			},
 			setupWanted: func() want {
@@ -106,14 +100,11 @@ func Test_getNodesHandle(t *testing.T) {
 		{
 			name: "retrieve nodes diff from a client that was already present",
 			setupArgs: func() args {
-				nodeDiscovery := nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil)
 				return args{
 					params: GetClusterNodesParams{
 						ClientID: &clientIDs[0],
 					},
-					daemon: &Daemon{
-						nodeDiscovery: nodeDiscovery,
-					},
+					nodeManager: g.nm,
 					clients: map[int64]*clusterNodesClient{
 						clientIDs[0]: {
 							ClusterNodeStatus: &models.ClusterNodeStatus{
@@ -159,14 +150,11 @@ func Test_getNodesHandle(t *testing.T) {
 		{
 			name: "retrieve nodes from an expired client, it should be ok because the clean up only happens when on insertion",
 			setupArgs: func() args {
-				nodeDiscovery := nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil)
 				return args{
 					params: GetClusterNodesParams{
 						ClientID: &clientIDs[0],
 					},
-					daemon: &Daemon{
-						nodeDiscovery: nodeDiscovery,
-					},
+					nodeManager: g.nm,
 					clients: map[int64]*clusterNodesClient{
 						clientIDs[0]: {
 							lastSync: time.Now().Add(-clientGCTimeout),
@@ -213,14 +201,11 @@ func Test_getNodesHandle(t *testing.T) {
 		{
 			name: "retrieve nodes for a new client, the expired client should be deleted",
 			setupArgs: func() args {
-				nodeDiscovery := nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil)
 				return args{
 					params: GetClusterNodesParams{
 						ClientID: &zero,
 					},
-					daemon: &Daemon{
-						nodeDiscovery: nodeDiscovery,
-					},
+					nodeManager: g.nm,
 					clients: map[int64]*clusterNodesClient{
 						clientIDs[numberOfClients-1]: {
 							lastSync: time.Now().Add(-clientGCTimeout),
@@ -262,14 +247,11 @@ func Test_getNodesHandle(t *testing.T) {
 		{
 			name: "retrieve nodes for a new client, however the randomizer allocated an existing clientID, so we should return a empty clientID",
 			setupArgs: func() args {
-				nodeDiscovery := nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil)
 				return args{
 					params: GetClusterNodesParams{
 						ClientID: &zero,
 					},
-					daemon: &Daemon{
-						nodeDiscovery: nodeDiscovery,
-					},
+					nodeManager: g.nm,
 					clients: map[int64]*clusterNodesClient{
 						clientIDs[0]: {
 							ClusterNodeStatus: &models.ClusterNodeStatus{
@@ -311,12 +293,9 @@ func Test_getNodesHandle(t *testing.T) {
 		{
 			name: "retrieve nodes for a client that does not want to have diffs, leave all other stored clients alone",
 			setupArgs: func() args {
-				nodeDiscovery := nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil)
 				return args{
-					params: GetClusterNodesParams{},
-					daemon: &Daemon{
-						nodeDiscovery: nodeDiscovery,
-					},
+					params:      GetClusterNodesParams{},
+					nodeManager: g.nm,
 					clients: map[int64]*clusterNodesClient{
 						clientIDs[0]: {
 							ClusterNodeStatus: &models.ClusterNodeStatus{
@@ -362,8 +341,11 @@ func Test_getNodesHandle(t *testing.T) {
 		randSrc.Seed(0, 0)
 		args := tt.setupArgs()
 		want := tt.setupWanted()
-		h := &getNodes{clients: args.clients}
-		responder := h.Handle(args.daemon, args.params)
+		h := &getClusterNodesRestApiHandler{
+			nodeManager: args.nodeManager,
+			clients:     args.clients,
+		}
+		responder := h.Handle(args.params)
 		require.EqualValues(t, len(want.clients), len(h.clients))
 		for k, v := range h.clients {
 			wantClient, ok := want.clients[k]
@@ -416,11 +398,11 @@ func Test_cleanupClients(t *testing.T) {
 		t.Log(tt.name)
 		args := tt.setupArgs()
 		want := tt.setupWanted()
-		h := &getNodes{clients: args.clients}
-		h.cleanupClients(
-			&Daemon{
-				nodeDiscovery: nodediscovery.NewNodeDiscovery(g.nm, nil, nil, &fake.FakeCNIConfigManager{}, nil),
-			})
+		h := &getClusterNodesRestApiHandler{
+			nodeManager: g.nm,
+			clients:     args.clients,
+		}
+		h.cleanupClients()
 		require.EqualValues(t, want.clients, h.clients)
 	}
 }
